Move log's typings into grafana/ui
