+++ public/app/features/explore/LogLabel.tsx
@@ -1,7 +1,8 @@
 import React, { PureComponent } from 'react';
 
-import { calculateLogsLabelStats, LogLabelStatsModel, LogRowModel } from 'app/core/logs_model';
 import { LogLabelStats } from './LogLabelStats';
+import { LogRowModel, LogLabelStatsModel } from '@grafana/ui';
+import { calculateLogsLabelStats } from 'app/core/logs_model';
 
 interface Props {
   getRows?: () => LogRowModel[];
