+++ public/app/features/explore/state/selectors.test.ts
@@ -1,5 +1,5 @@
 import { deduplicatedLogsSelector } from './selectors';
-import { LogsDedupStrategy } from 'app/core/logs_model';
+import { LogsDedupStrategy } from '@grafana/ui';
 import { ExploreItemState } from 'app/types';
 
 const state = {
